move types from `methodIds.ts` to `transactions.d.ts`
